# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "CarRent Support",
            "url": "https://www.carrent.com/support",
            "email": "support@carrent.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/role": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get roles data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Get roles data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search by id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by scope",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search by client_id",
                        "name": "cid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort result by attributes",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": " ",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RolesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.RolesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.RolesResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create role data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Create Role",
                "parameters": [
                    {
                        "description": "Role Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    }
                }
            }
        },
        "/role/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get roles data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Get roles data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "must-revalidate",
                            "none"
                        ],
                        "type": "string",
                        "description": "Request Cache Control",
                        "name": "Cache-Control",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Update role data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Update role data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateRole"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.SingleRoleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Delete role data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "Delete role data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.EmptyResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.CreateRole": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                }
            }
        },
        "model.EmptyResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
...
        "model.Pagination": {
            "type": "object",
            "properties": {
                "current_elements": {
                    "type": "integer"
                },
                "current_page": {
                    "type": "integer"
                },
                "sort_by": {
                    "type": "string"
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "scope": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
        "model.RolesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.Pagination"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
...
        "model.SingleRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Role"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "transaction_info": {
                    "$ref": "#/definitions/model.TransactionInfo"
                },
                "translation": {
                    "$ref": "#/definitions/model.Translation"
                }
            }
        },
        "model.TransactionInfo": {
            "type": "object",
            "properties": {
                "cause": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "string"
                },
                "request_method": {
                    "type": "string"
                },
                "request_uri": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "model.Translation": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateRole": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "%s"
        }
    }
}